// types.ts
export type CurrencyCode = 'USD' | 'IDR';